live_trading/requirements.txt
# Core dependencies for live trading bot
# Install with: pip install -r requirements.txt

# HTTP requests for API calls
requests==2.31.0

# Environment variable management
python-dotenv==1.0.0

# Timezone handling
pytz==2024.1

# Testing dependencies (optional)
pytest==7.4.3
pytest-cov==4.1.0

# Development dependencies (optional)
black==23.12.1  # Code formatter
flake8==7.0.0   # Linter
